requirements.txt
# DAWO.ECO Dependencies
# Generated by Story 3-2-google-drive-asset-storage

# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
redis>=5.0.0
arq>=0.25.0

# HTTP Client
httpx>=0.26.0

# Google APIs (Story 3.2)
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0

# Google Gemini AI (Story 3.5)
google-generativeai>=0.8.0

# Image Processing (Story 3.5)
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.8.0
